move modern FeeLabel
